backend/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
mysqlclient==1.3.12
django-filter==1.0.4
requests==2.18.4
django-cors-headers==2.1.0